+++ db/sqlc/store.go
@@ -4,14 +4,21 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"github.com/google/uuid"
 )
 
-// Store provides all functions to execute db queries and transactions
+// Store provides all functions to execute database queries and transactions.
+// It embeds *Queries to allow direct access to query methods and maintains a reference to the database connection.
 type Store struct {
 	*Queries
 	db *sql.DB
 }
 
+// NewStore creates a new Store instance.
+// Parameters:
+// - db: A pointer to the database connection.
+// Returns:
+// - A pointer to the initialized Store.
 func NewStore(db *sql.DB) *Store {
 	return &Store{
 		db:      db,
@@ -19,9 +26,14 @@ func NewStore(db *sql.DB) *Store {
 	}
 }
 
+// execTx executes a function within a database transaction.
+// Parameters:
+// - ctx: The context for the transaction.
+// - fn: A function that takes *Queries and performs database operations.
+// Returns:
+// - An error if the transaction fails or the function returns an error.
 func (store *Store) execTx(ctx context.Context, fn func(*Queries) error) error {
 	tx, err := store.db.BeginTx(ctx, nil)
-
 	if err != nil {
 		return err
 	}
@@ -41,14 +53,26 @@ func (store *Store) execTx(ctx context.Context, fn func(*Queries) error) error {
 	return tx.Commit()
 }
 
-// UserTxResult CreateUserTxResult is the result of transfer transaction.
+// UserTxResult represents the result of a transactional operation involving a user.
+// Fields:
+// - User: The user entity.
+// - UserProfile: The associated user profile entity.
+// - UserRole: The associated user role entity.
 type UserTxResult struct {
 	User        User        `json:"user"`
 	UserProfile UserProfile `json:"user_profile"`
 	UserRole    UserRole    `json:"user_role"`
 }
 
-// CreateUserWithProfileAndRoleTx performs a transaction to create a new user entry along with their profile and role information all in one go.
+// CreateUserWithProfileAndRoleTx performs a transaction to create a user, their profile, and role.
+// Parameters:
+// - ctx: The context for the transaction.
+// - userParams: Parameters for creating the user.
+// - profileParams: Parameters for creating the user profile.
+// - roleParams: Parameters for creating the user role.
+// Returns:
+// - A UserTxResult containing the created user, profile, and role.
+// - An error if the transaction fails.
 func (store *Store) CreateUserWithProfileAndRoleTx(ctx context.Context, userParams CreateUserParams, profileParams CreateUserProfileParams, roleParams CreateUserRoleParams) (UserTxResult, error) {
 	var result UserTxResult
 
@@ -56,7 +80,6 @@ func (store *Store) CreateUserWithProfileAndRoleTx(ctx context.Context, userPara
 		var err error
 
 		result.User, err = q.CreateUser(ctx, userParams)
-
 		if err != nil {
 			return err
 		}
@@ -64,7 +87,6 @@ func (store *Store) CreateUserWithProfileAndRoleTx(ctx context.Context, userPara
 		profileParams.UserID = result.User.ID
 
 		result.UserProfile, err = q.CreateUserProfile(ctx, profileParams)
-
 		if err != nil {
 			return err
 		}
@@ -72,7 +94,112 @@ func (store *Store) CreateUserWithProfileAndRoleTx(ctx context.Context, userPara
 		roleParams.UserID = result.User.ID
 
 		result.UserRole, err = q.CreateUserRole(ctx, roleParams)
+		if err != nil {
+			return err
+		}
 
+		return nil
+	})
+
+	return result, err
+}
+
+// GetUserWithProfileAndRoleTX retrieves a user, their profile, and role in a single transaction.
+// Parameters:
+// - ctx: The context for the transaction.
+// - userID: The UUID of the user to retrieve.
+// Returns:
+// - A UserTxResult containing the user, profile, and role.
+// - An error if the transaction fails or the user does not exist.
+func (store *Store) GetUserWithProfileAndRoleTX(ctx context.Context, userID uuid.UUID) (UserTxResult, error) {
+	var result UserTxResult
+	err := store.execTx(ctx, func(q *Queries) error {
+		var err error
+
+		result.User, err = q.GetUser(ctx, userID)
+		if err != nil {
+			return err
+		}
+
+		result.UserProfile, err = q.GetUserProfile(ctx, userID)
+		if err != nil {
+			return err
+		}
+
+		result.UserRole, err = q.GetUserRole(ctx, userID)
+		if err != nil {
+			return err
+		}
+
+		return nil
+	})
+
+	return result, err
+}
+
+// DeleteUserWithProfileAndRoleTX deletes a user, their profile, and role in a single transaction.
+// Parameters:
+// - ctx: The context for the transaction.
+// - userID: The UUID of the user to delete.
+// Returns:
+// - A UserTxResult containing the deleted user, profile, and role.
+// - An error if the transaction fails or the user does not exist.
+func (store *Store) DeleteUserWithProfileAndRoleTX(ctx context.Context, userID uuid.UUID) (UserTxResult, error) {
+	var result UserTxResult
+
+	err := store.execTx(ctx, func(q *Queries) error {
+		var err error
+
+		result.UserRole, err = q.DeleteUserRole(ctx, userID)
+		if err != nil {
+			return err
+		}
+
+		result.UserProfile, err = q.DeleteUserProfile(ctx, userID)
+		if err != nil {
+			return err
+		}
+
+		result.User, err = q.DeleteUser(ctx, userID)
+		if err != nil {
+			return err
+		}
+		return nil
+	})
+
+	return result, err
+}
+
+// UpdateUserWithProfileAndRoleTX updates a user, their profile, and role in a single transaction.
+// Parameters:
+// - ctx: The context for the transaction.
+// - userParams: Parameters for updating the user.
+// - profileParams: Parameters for updating the user profile.
+// - roleParams: Parameters for updating the user role.
+// Returns:
+// - A UserTxResult containing the updated user, profile, and role.
+// - An error if the transaction fails or the user does not exist.
+func (store *Store) UpdateUserWithProfileAndRoleTX(ctx context.Context, userParams UpdateUserParams, profileParams UpdateUserProfileParams, roleParams UpdateUserRoleParams) (UserTxResult, error) {
+	var result UserTxResult
+
+	err := store.execTx(ctx, func(q *Queries) error {
+		var err error
+
+		result.User, err = q.UpdateUser(ctx, userParams)
+		if err != nil {
+			return err
+		}
+
+		profileParams.UserID = userParams.ID
+
+		result.UserProfile, err = q.UpdateUserProfile(ctx, profileParams)
+		if err != nil {
+			return err
+		}
+
+		roleParams.UserID = userParams.ID
+
+		result.UserRole, err = q.UpdateUserRole(ctx, roleParams)
 		if err != nil {
 			return err
 		}