+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Next Release
 -
 
+v1.4.0
+* Install Gin v1.6.3
+* Setup Create User Route
+* Setup Get User Route
+* Setup List User Route
+* Setup Delete User Route
+* Setup Update User Route
+* Get User Transaction and Tests
+* Delete User Transaction and Tests
+* Update User Transaction and Tests
+* Setup Create User Transaction Route
+* Setup Get User Transaction Route
+* Setup Delete User Transaction Route
+* Setup Update User Transaction Route
+
 v1.3.0
 * Adding CI Testing
 * Add Transaction Setup